refactor(runtime): move authtest under runtime/internal (archtest B → Go compiler Hard)
The runtime/auth/authtest package (PR #267 substrate) was a top-level test-only
helper whose access boundary was enforced by archtest AUTH-AUTHTEST-B (a
string-anchor scan over cells/**, examples/**, kernel/** imports). Per
.claude/rules/gocell/ai-robust.md §Hard 范本目录 → "internal/ wrap 包", this
moves the package to runtime/internal/authtest/ so the Go compiler refuses
out-of-runtime/ imports at compile time. AUTH-AUTHTEST-B is retired as
redundant; A (literal call ban) and C (only _test.go can import) remain as
archtest Medium with documented terminal rationale.

Path selection: runtime/internal/authtest/ (NOT runtime/auth/internal/authtest/
as the original backlog entry proposed) — verified consumers span
runtime/bootstrap, runtime/http/router, both outside runtime/auth/**, so the
auth-internal placement would break their imports.

ref: golang/go cmd/go/#hdr-Internal_Directories
ref: kubernetes-sigs/controller-runtime — internal/ subtree visibility pattern

Closes #638
