+++ runtime/http/router/router_authmeta_test.go
@@ -17,7 +17,7 @@ import (
 	kcell "github.com/ghbvf/gocell/kernel/cell"
 	"github.com/ghbvf/gocell/kernel/clock"
 	"github.com/ghbvf/gocell/runtime/auth"
-	"github.com/ghbvf/gocell/runtime/auth/authtest"
+	"github.com/ghbvf/gocell/runtime/internal/authtest"
 )
 
 // ---------------------------------------------------------------------------
